fix: export suppressLogs
fix: config global when createStandardLog is called the first time
