+++ .changeset/dirty-cherries-visit.md
@@ -1,0 +1,8 @@
+---
+"standard-log": patch
+---
+
+Exports `suppressLogs()`.
+
+The global instance will be configured when `createStandardLog()` is called,
+if `configGlobal()` is not called before.
