+++ packages/log/ts/index.ts
@@ -1,11 +1,11 @@
 export * from './console.js'
 export * from './errors.js'
 export * from './formatter.js'
-export { configGlobal, getLogger } from './global.js'
 export * from './logLevels.js'
 export * from './memory.js'
 export {
-  createStandardLog, createStandardLogForTest, StandardLog, StandardLogInstance
+  configGlobal, createStandardLog, createStandardLogForTest,
+  getLogger, StandardLog, StandardLogInstance, suppressLogs
 } from './standardLog.js'
 export * from './types.js'
