+++ packages/log/ts/standardLog.ts
@@ -1,5 +1,5 @@
 import { StackTraceMeta } from '@just-func/types'
-import { required } from 'type-plus'
+import { record, required } from 'type-plus'
 import { createLogger } from './logger.js'
 import { logLevels } from './logLevels.js'
 import { createLogStore, LogStore } from './logStore.js'
@@ -19,7 +19,9 @@ export interface StandardLog<N extends string = LogMethodNames> extends Readonly
 export function createStandardLog<N extends string = LogMethodNames>(
   options?: StandardLogOptions<N>
 ): StandardLog<N> {
-  return createStandardLogClosure(options).standardLog
+  const closure = createStandardLogClosure(options)
+  if (!ctx.gsl) ctx.gsl = closure as any
+  return closure.standardLog
 }
 
 export function createStandardLogClosure<N extends string = LogMethodNames>(
@@ -62,4 +64,42 @@ export function suppressLogs<R>(block: () => R, ...logs: Logger[]): R {
   const result = block()
   logs.forEach((l, i) => l.level = origLevels[i])
   return result
+}
+
+export function getLogger<N extends string = LogMethodNames>(
+  params: [id: string, options?: LoggerOptions], meta?: StackTraceMeta) {
+  return getGlobalSL<N>().getLogger(params, meta)
+}
+
+export const ctx: {
+  configured?: boolean,
+  gsl?: { store: LogStore, standardLog: StandardLogInstance<LogMethodNames> }
+} = record()
+
+function getGlobalSL<N extends string = LogMethodNames>(): StandardLogInstance<N> {
+  if (!ctx.gsl) ctx.gsl = createStandardLogClosure()
+
+  return ctx.gsl!.standardLog as StandardLogInstance<N>
+}
+
+export function configGlobal(options: Omit<StandardLogOptions, 'customLevels'>) {
+  if (ctx.configured) {
+    const log = ctx.gsl?.standardLog.getLogger(['standard-log'])
+    log?.warn('configGlobal() is being called more than once. Please make sure this is expected. Application should use `createStandardLog()` most of the time.')
+  }
+
+  if (ctx.gsl) {
+    if (options.logLevel) {
+      ctx.gsl.store.logLevel = options.logLevel
+    }
+    if (options.reporters) {
+      const reporters = ctx.gsl.store.reporters
+      while (reporters.length > 0) reporters.pop()
+      reporters.push(...options.reporters)
+    }
+  }
+  else {
+    ctx.gsl = createStandardLogClosure(options)
+  }
+  ctx.configured = true
 }
